Moves "augment" to experimental
This is an intresting decorator, but might not end up making it into
the final API. That said, we can use it to test the decorator framework,
and this will give us a sense of whether users are acutally intereted in
it and what could change.
